requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.10
sentence-transformers==5.2.2
faiss-cpu==1.7.4
pypdf2==3.0.1
cryptography==41.0.7
werkzeug==3.0.1
gunicorn==21.2.0
